% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.R
\name{gds}
\alias{gds}
\title{Descriptive statistics of grouped data: with the help of this package we calculate mean, median, mode, variance, standard deviation, coefficient of variance, quartiles, IQR, skewness, and kurtosis of grouped data.}
\usage{
gds(ll, ul, freq)
}
\arguments{
\item{ll}{A data vector to store lower limit of the classes}

\item{ul}{A data vector to store upper limit of the classes}

\item{freq}{A data vector to store the frequencies of the corresponding classes}
}
\value{
gmean, gmedian, gmode, gvar, gstdev, gcv, gq1, gq2, gq3, gIQR, g1, g2
}
\description{
Descriptive statistics of grouped data: with the help of this package we calculate mean, median, mode, variance, standard deviation, coefficient of variance, quartiles, IQR, skewness, and kurtosis of grouped data.
}
\examples{
gds(c(10,20,30,40,50),c(20,30,40,50,60),c(7,13,23,20,8))


}
\references{
1. Gupta, S.P., and Gupta, M.P. (2005) Business statistics, Sultan Chand and Sons educational publishers, New Delhi.

2. Levine, D.M., Krehbiel,  T.C.,  Bereson, M.L. and Viswanathan, P.K. (2011) Business statistics: a first course, 5th edition, Pearson.

3. Langford, E. (2006) Quartiles in Elementary Statistics, Journal of Statistics Education Volume 14, Number 3.

4. Das, N. G. (2010) Statistical Methods- Combined Edition (Volumes I & II), Tata McGraw Hill Education Private Limited, New Delhi.
}
